@hathach Arduino core now wants size_t returned

+++ cores/arduino/Adafruit_TinyUSB_Core/Adafruit_USBD_CDC.h
@@ -51,7 +51,7 @@ public:
 	size_t write(const char *buffer, size_t size) {
 	  return write((const uint8_t *)buffer, size);
 	}
-	virtual int availableForWrite(void);
+	virtual size_t availableForWrite(void);
 	operator bool();
 };
